+++ src/undated/__init__.py
@@ -8,7 +8,7 @@ All of these are hopefully self explanitory.
 
 from ._tools import *
 
-__version__ = '1.0.7'
+__version__ = '1.0.8'
 
 # -----------------------------------------------
 # End.