+++ docs/changes.rst
@@ -1,2 +1,10 @@
 Change Log
 ==========
+
+Version 1.0.8
+-------------
+*Date* 10th October 2022
+
+Added new functions to utils, `first_day` and `last_day`.
+These convert year month integers to year month day integers,
+with the day element being either the first or last day of the month.